Move data parsing to App component
